+++ src/main/java/com/seoultech/synergybe/domain/auth/controller/AuthController.java
@@ -1,11 +1,11 @@
-package com.seoultech.synergybe.domain.auth.api;
+package com.seoultech.synergybe.domain.auth.controller;
 
 import com.seoultech.synergybe.domain.auth.entity.AuthReqModel;
 import com.seoultech.synergybe.domain.auth.entity.RoleType;
 import com.seoultech.synergybe.domain.auth.entity.UserPrincipal;
 import com.seoultech.synergybe.domain.auth.token.AuthToken;
 import com.seoultech.synergybe.domain.auth.token.AuthTokenProvider;
-import com.seoultech.synergybe.domain.user.dao.UserRefreshTokenRepository;
+import com.seoultech.synergybe.domain.user.repository.UserRefreshTokenRepository;
 import com.seoultech.synergybe.domain.user.entity.UserRefreshToken;
 import com.seoultech.synergybe.system.common.ApiResponse;
 import com.seoultech.synergybe.system.config.properties.AppProperties;
