+++ src/main/java/com/seoultech/synergybe/domain/auth/service/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
-package com.seoultech.synergybe.domain.auth.application;
+package com.seoultech.synergybe.domain.auth.service;
 
 import com.seoultech.synergybe.domain.auth.entity.UserPrincipal;
-import com.seoultech.synergybe.domain.user.dao.UserRepository;
+import com.seoultech.synergybe.domain.user.repository.UserRepository;
 import com.seoultech.synergybe.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
