+++ src/main/java/com/seoultech/synergybe/system/config/security/SecurityConfig.java
@@ -1,11 +1,11 @@
 package com.seoultech.synergybe.system.config.security;
 
-import com.seoultech.synergybe.domain.auth.application.CustomOAuth2UserService;
-import com.seoultech.synergybe.domain.auth.application.CustomUserDetailsService;
-import com.seoultech.synergybe.domain.auth.dao.OAuth2AuthorizationRequestBasedOnCookieRepository;
+import com.seoultech.synergybe.domain.auth.service.CustomOAuth2UserService;
+import com.seoultech.synergybe.domain.auth.service.CustomUserDetailsService;
+import com.seoultech.synergybe.domain.auth.repository.OAuth2AuthorizationRequestBasedOnCookieRepository;
 import com.seoultech.synergybe.domain.auth.entity.RoleType;
 import com.seoultech.synergybe.domain.auth.token.AuthTokenProvider;
-import com.seoultech.synergybe.domain.user.dao.UserRefreshTokenRepository;
+import com.seoultech.synergybe.domain.user.repository.UserRefreshTokenRepository;
 import com.seoultech.synergybe.system.config.properties.AppProperties;
 import com.seoultech.synergybe.system.config.properties.CorsProperties;
 import com.seoultech.synergybe.system.exception.RestAuthenticationEntryPoint;
@@ -16,14 +16,10 @@ import com.seoultech.synergybe.system.handler.TokenAccessDeniedHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.BeanIds;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.config.ldap.LdapBindAuthenticationManagerFactory;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
